prediction_service/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
PyYAML>=6.0

# Cassandra driver
cassandra-driver>=3.28.0

# Technical analysis
ta>=0.10.2

# Web framework for health check
Flask>=2.3.0
Werkzeug>=2.3.0

# Logging and utilities
python-dateutil>=2.8.0
tqdm>=4.64.0

# Math and data processing
matplotlib>=3.6.0
seaborn>=0.12.0

# Optional: for better performance.
# xformers  # Uncomment if using FlashAttention model.